refactor: migrate provinces-editor to bundled TypeScript controller
Port public/modules/ui/provinces-editor.js (1372 lines) to
src/controllers/provinces-editor.ts following the migration guide (strict
d3 v7: select/drag/pointer/stratify/treemap/transition/color/interpolate
named imports, event-arg brush and treemap handlers). The static
#provincesEditor chrome stays in index.html (rows generated at runtime);
listeners wire once behind initialized flags. Register as
Controllers.ProvincesEditor and replace editProvinces() at the tools.js and
hotkeys.js call sites. Extend the Province interface with the editor's
computed stat fields (area/rural/urban/burgs).
