+++ src/controllers/index.ts
@@ -31,6 +31,7 @@ export const Controllers = createRegistry({
   NotesEditor: () => import("@/controllers/notes-editor").then(m => m.NotesEditor),
   ProductionChains: () => import("@/controllers/production-chains").then(m => m.ProductionChains),
   ProductionOverview: () => import("@/controllers/production-overview").then(m => m.ProductionOverview),
+  ProvincesEditor: () => import("@/controllers/provinces-editor").then(m => m.ProvincesEditor),
   RegimentEditor: () => import("@/controllers/regiment-editor").then(m => m.RegimentEditor),
   RegimentsOverview: () => import("@/controllers/regiments-overview").then(m => m.RegimentsOverview),
   ReliefEditor: () => import("@/controllers/relief-editor").then(m => m.ReliefEditor),
